Move all the files that contain app commands into their own dirs
Create a deployer/commands directory and put all of the code that
defines deployer commands under this directory, including the main
`deployer.py` file.
