+++ deployer/commands/exec/shell/app.py
@@ -5,7 +5,7 @@ under the `exec` sub-command of the deployer.
 """
 import typer
 
-from ...cli_app import exec_app
+from ....cli_app import exec_app
 
 exec_shell_app = typer.Typer(pretty_exceptions_show_locals=False)
 exec_app.add_typer(exec_shell_app, name="shell")
