+++ deployer/commands/exec/shell/infra_components_commands.py
@@ -4,10 +4,8 @@ import subprocess
 import typer
 from ruamel.yaml import YAML
 
-from ...cluster import Cluster
-from ...file_acquisition import find_absolute_path_to_cluster_file
-from ...utils import print_colour
-
+from ....cluster import Cluster
+from ....file_acquisition import find_absolute_path_to_cluster_file
 from .app import exec_shell_app
 
 # Without `pure=True`, I get an exception about str / byte issues
