+++ deployer/commands/generate/billing/app_and_commands.py
@@ -5,8 +5,8 @@ import typer
 from dateutil.relativedelta import relativedelta
 from ruamel.yaml import YAML
 
-from ...cli_app import generate_app
-from ...helm_upgrade_decision import get_all_cluster_yaml_files
+from ....cli_app import generate_app
+from ....helm_upgrade_decision import get_all_cluster_yaml_files
 from .importers import get_cluster_costs
 from .outputers import CostTableOutputFormats, output_cost_table
