+++ deployer/commands/grafana/deploy_dashboards.py
@@ -4,8 +4,8 @@ import subprocess
 
 import typer
 
-from ..cli_app import grafana_app
-from ..utils import print_colour
+from ...cli_app import grafana_app
+from ...utils import print_colour
 from .utils import get_grafana_token, get_grafana_url
