+++ deployer/commands/grafana/tokens.py
@@ -15,8 +15,8 @@ from pathlib import Path
 import requests
 import typer
 
-from ..cli_app import grafana_app
-from ..utils import print_colour
+from ...cli_app import grafana_app
+from ...utils import print_colour
 from .utils import (
     get_grafana_admin_password,
     get_grafana_url,
