+++ deployer/commands/grafana/utils.py
@@ -5,7 +5,7 @@ from pathlib import Path
 import requests
 from ruamel.yaml import YAML
 
-from ..file_acquisition import find_absolute_path_to_cluster_file, get_decrypted_file
+from ...file_acquisition import find_absolute_path_to_cluster_file, get_decrypted_file
 
 yaml = YAML(typ="safe")
